finish frontend

+++ zefzafy-blog-backend/src/category/category.service.ts
@@ -32,8 +32,33 @@ export class CategoryService {
     return this.categoryRepository.save(category);
   }
 
-  public async findAll() {
-    return this.categoryRepository.find();
+  public async findAll(page: number, limit: number) {
+    const pageNumber = Math.max(1, page);
+    const limitNumber = Math.max(1, limit);
+
+    // Calculate skip (offset) for pagination
+    const skip = (pageNumber - 1) * limitNumber;
+
+    // Fetch paginated users and total count
+    const [categories, total] = await this.categoryRepository.findAndCount({
+      skip,
+      take: limitNumber,
+      relations : {posts : true}
+    });
+
+    // Calculate total pages
+    const pagesCount = Math.ceil(total / limitNumber);
+
+    // Return response in desired format
+    return {
+      categories,
+      pagination: {
+        total,
+        page: pageNumber,
+        limit: limitNumber,
+        pagesCount,
+      },
+    };
   }
 
   public async findOne(id: number) {
@@ -57,21 +82,29 @@ export class CategoryService {
   public async remove(id: number) {
     const category = await this.findOne(id);
     if (!category) throw new NotFoundException('category not found');
- 
-   const posts = await this.postService.findCtegoryPosts(id);
 
-   let publicIds =[];
-   posts.map(post => {
-    if (post.image) {
-      publicIds.push(post.image.public_id) ;
+    const posts = await this.postService.findCtegoryPosts(id);
+
+    let publicIds = [];
+    if (posts.length > 0) {
+    posts.map((post) => {
+      if (post.image) {
+        publicIds.push(post.image.public_id);
+      }
+    });
     }
-   });
-   console.log(publicIds);
-   
-   //TODO: remove images from cloudinary
-   await this.categoryRepository.remove(category);
-   
-   await this.cloudinaryService.removeMultipleImages(publicIds);
+    if (publicIds.length > 0 ) {
+      await this.cloudinaryService.removeMultipleImages(publicIds);
+    }
+
+
+    await this.categoryRepository.remove(category);
+
+  
     return { message: `category with id (${id}) was removed` };
+  }
+
+  async getCategoriesCount() {
+    return this.categoryRepository.count();
   }
 }
